add commands.cog
