+++ data/accounts/awexander#007.json
@@ -95,6 +95,48 @@
         "matchid":"dc79ec1f-be86-4095-8a88-ad275d8a175c",
         "score":"4-13",
         "agent":"Chamber",
+        "headshot":16,
+        "kda":[
+            21,
+            14,
+            2,
+            1.5
+        ],
+        "adr":212
+    },
+    {
+        "account":{
+            "name":"awexander",
+            "tag":"007"
+        },
+        "rank":"Diamond 1",
+        "map":"Breeze",
+        "mode":"Competitive",
+        "timeplayed":"August 20, 2022 at 22:12 GMT+8",
+        "matchid":"dc79ec1f-be86-4095-8a88-ad275d8a175c",
+        "score":"4-13",
+        "agent":"Chamber",
+        "headshot":16,
+        "kda":[
+            21,
+            14,
+            2,
+            1.5
+        ],
+        "adr":212
+    },
+    {
+        "account":{
+            "name":"awexander",
+            "tag":"007"
+        },
+        "rank":"Diamond 1",
+        "map":"Breeze",
+        "mode":"Competitive",
+        "timeplayed":"August 20, 2022 at 22:12 GMT+8",
+        "matchid":"dc79ec1f-be86-4095-8a88-ad275d8a175c",
+        "score":"4-13",
+        "agent":"Chamber",
         "headshot":0,
         "kda":[
             21,
