+++ data/accounts/belangsukawa#1001.json
@@ -116,6 +116,48 @@
         "matchid":"64054e1d-aa40-4261-9a53-8b7c26b7ed4a",
         "score":"9-13",
         "agent":"Sage",
+        "headshot":24,
+        "kda":[
+            18,
+            14,
+            7,
+            1.3
+        ],
+        "adr":143
+    },
+    {
+        "account":{
+            "name":"belangsukawa",
+            "tag":"1001"
+        },
+        "rank":"Platinum 2",
+        "map":"Breeze",
+        "mode":"Competitive",
+        "timeplayed":"August 23, 2022 at 21:27 GMT+8",
+        "matchid":"64054e1d-aa40-4261-9a53-8b7c26b7ed4a",
+        "score":"9-13",
+        "agent":"Sage",
+        "headshot":24,
+        "kda":[
+            18,
+            14,
+            7,
+            1.3
+        ],
+        "adr":143
+    },
+    {
+        "account":{
+            "name":"belangsukawa",
+            "tag":"1001"
+        },
+        "rank":"Platinum 2",
+        "map":"Breeze",
+        "mode":"Competitive",
+        "timeplayed":"August 23, 2022 at 21:27 GMT+8",
+        "matchid":"64054e1d-aa40-4261-9a53-8b7c26b7ed4a",
+        "score":"9-13",
+        "agent":"Sage",
         "headshot":0,
         "kda":[
             18,
